+++ x/fake/writer_test.go
@@ -105,14 +105,14 @@ func TestWrite_Compound(t *testing.T) {
 	fakeData1 := []byte("F")
 	fakeBytes1 := int64(1)
 	fakeOffset1 := int64(0)
-	fakeTtl1 := int64(0)
+	fakeTtl1 := 0
 	writer1 := NewWriter(&innerWriter, fakeBytes1, fakeData1, fakeOffset1, fakeTtl1)
 
 	// Second fakeWriter: fakeBytes=3, fakeData="ake d", fakeOffset=0
 	fakeData2 := []byte("ake") // Total fakeData now: "Fake d"
 	fakeBytes2 := int64(3)
 	fakeOffset2 := int64(0)
-	fakeTtl2 := int64(0)
+	fakeTtl2 := 0
 	fakeWriter := NewWriter(writer1, fakeBytes2, fakeData2, fakeOffset2, fakeTtl2)
 
 	// Write "Request"
@@ -205,14 +205,14 @@ func TestReadFrom_Compound(t *testing.T) {
 	fakeData1 := []byte("Fake ")
 	fakeBytes1 := int64(3)
 	fakeOffset1 := int64(0)
-	fakeTtl1 := int64(0)
+	fakeTtl1 := 0
 	writer1 := NewWriter(&innerWriter, fakeBytes1, fakeData1, fakeOffset1, fakeTtl1)
 
 	// Second fakeWriter: fakeBytes=5, fakeData="data", fakeOffset=0
 	fakeData2 := []byte("data")
 	fakeBytes2 := int64(5)
 	fakeOffset2 := int64(0)
-	fakeTtl2 := int64(0)
+	fakeTtl2 := 0
 	writer2 := NewWriter(writer1, fakeBytes2, fakeData2, fakeOffset2, fakeTtl2)
 
 	n, err := writer2.Write([]byte("Request"))
